feat: change theme color to blue to match travel branding
Originally, green was chosen as the branding color to represent money
management.

Blue was chosen in place to give users the idea of travel and leisure,
and the open skies. Avoiding users to be stressed out by finances is
better. This made more sense from a branding and marketing perspective.